requirements.txt
asgiref==3.7.2
autopep8==2.0.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Django==4.2.3
djangorestframework==3.14.0
docopt==0.6.2
flake8==6.0.0
ghp-import==2.1.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.5.1
mkdocs-material==9.1.21
mkdocs-material-extensions==1.1.1
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
pymdown-extensions==10.1
pytest==7.4.0
pytest-django==4.5.2
pytest-watch==4.2.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2023.6.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
watchdog==3.0.0
